Still on dtype

+++ losses.py
@@ -17,8 +17,8 @@ class log_loss():
 
     def grad(X, y, w):
         return torch.mean(torch.div(-torch.mul(X.t(), y.double()),
-                          (1. + torch.exp(y * torch.sum(torch.mul(X, w))))),
-                          dim=1)
+                          (1. + torch.exp(y.double() * torch.sum(torch.mul(X,
+                           w))))), dim=1)
 
 
 class prox_loss():
